Refactored to use single dashboard
