--- src/routes/views/Overview/BillingDashboard/BillingDashboardWidget.styles.ts
@@ -1,4 +1,0 @@
-export const chartStyles = {
-  chartAltHeight: 250,
-  containerAltHeight: 250,
-};
