--- src/routes/views/Overview/BillingDashboard/index.ts
@@ -1,1 +1,0 @@
-export { default as BillingDashboard } from './BillingDashboard';
