+++ src/routes/views/Overview/components/Dashboard/Dashboard.tsx
@@ -1,12 +1,12 @@
 import { Grid, GridItem } from '@patternfly/react-core';
 import React from 'react';
-import { WrappedComponentProps } from 'react-intl';
-import { BillingCardSize } from 'store/dashboard/billingDashboard';
+import { connect } from 'react-redux';
+import { createMapStateToProps } from 'store/common';
+import { dashboardSelectors, DashboardSize } from 'store/dashboard';
 
-type DashboardOwnProps = WrappedComponentProps;
+type DashboardOwnProps = any;
 
 interface DashboardStateProps {
-  DashboardWidget: any;
   widgets: number[];
 }
 
@@ -20,7 +20,7 @@ const DashboardBase: React.SFC<DashboardProps> = ({ selectWidgets, widgets }) =>
   <Grid hasGutter>
     {widgets.map(widgetId => {
       const widget = selectWidgets[widgetId];
-      return widget.size === BillingCardSize.full ? (
+      return widget.size === DashboardSize.full ? (
         <GridItem sm={12} key={widgetId}>
           <widget.component widgetId={widgetId} />
         </GridItem>
@@ -33,4 +33,14 @@ const DashboardBase: React.SFC<DashboardProps> = ({ selectWidgets, widgets }) =>
   </Grid>
 );
 
-export default DashboardBase;
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+const mapStateToProps = createMapStateToProps<DashboardOwnProps, DashboardStateProps>((state, props) => {
+  return {
+    selectWidgets: dashboardSelectors.selectWidgets(state),
+    widgets: dashboardSelectors.selectCurrentWidgets(state),
+  };
+});
+
+const Dashboard = connect(mapStateToProps, {})(DashboardBase);
+
+export default Dashboard;
