+++ src/routes/views/Overview/components/Dashboard/index.ts
@@ -1,7 +1,1 @@
-export { default as DashboardBase } from './DashboardBase';
-export {
-  default as DashboardWidgetBase,
-  DashboardWidgetOwnProps,
-  DashboardWidgetProps,
-  DashboardWidgetStateProps,
-} from './DashboardWidgetBase';
+export { default as Dashboard } from './Dashboard';
