--- src/store/dashboard/billingDashboard/index.ts
@@ -1,13 +1,0 @@
-import * as billingDashboardActions from './billingDashboardActions';
-import { BillingCardType, billingDashboardStateKey, BillingDashboardWidget } from './billingDashboardCommon';
-import { billingDashboardReducer } from './billingDashboardReducer';
-import * as billingDashboardSelectors from './billingDashboardSelectors';
-
-export {
-  billingDashboardStateKey,
-  billingDashboardReducer,
-  billingDashboardActions,
-  billingDashboardSelectors,
-  BillingCardType,
-  BillingDashboardWidget,
-};
