+++ src/store/dashboard/index.ts
@@ -1,0 +1,5 @@
+import { DashboardSize, dashboardStateKey, DashboardType, DashboardWidget } from './dashboardCommon';
+import { dashboardReducer } from './dashboardReducer';
+import * as dashboardSelectors from './dashboardSelectors';
+
+export { dashboardStateKey, dashboardReducer, dashboardSelectors, DashboardType, DashboardSize, DashboardWidget };
