+++ src/store/rootReducer.ts
@@ -1,6 +1,6 @@
 import { notificationsReducer } from '@redhat-cloud-services/frontend-components-notifications/redux';
 import { combineReducers } from 'redux';
-import { billingDashboardReducer, billingDashboardStateKey } from 'store/dashboard/billingDashboard';
+import { dashboardReducer, dashboardStateKey } from 'store/dashboard';
 import { reportReducer, reportStateKey } from 'store/reports';
 import { userAccessReducer, userAccessStateKey } from 'store/userAccess';
 import { StateType } from 'typesafe-actions';
@@ -9,7 +9,7 @@ import { featureFlagsReducer, featureFlagsStateKey } from './featureFlags';
 export type RootState = StateType<typeof rootReducer>;
 
 export const rootReducer = combineReducers({
-  [billingDashboardStateKey]: billingDashboardReducer,
+  [dashboardStateKey]: dashboardReducer,
   [featureFlagsStateKey]: featureFlagsReducer,
   [reportStateKey]: reportReducer,
   [userAccessStateKey]: userAccessReducer,
